added video background for homepage
