+++ README.md
@@ -3,3 +3,6 @@ refactoring a project to MERN application
 
 
 ## Technologies used
+
+## Images & Videos
+Beach- https://mixkit.co/free-stock-video/beach-seen-from-the-air-with-waves-reaching-the-shore-51461/  Mixkit Stock Video Free License.
