+++ client/src/App.jsx
@@ -12,9 +12,9 @@ function App() {
       <div>
         <Navbar />
         <Routes>
-          <Route exact="true" path="/" component= {Home} />
-          <Route path="/itinerary" component={Itinerary} />
-          <Route path="/profile" component={Profile} />
+          <Route exact path ="/" element= {<Home />} />
+          <Route path="/itinerary" element= {<Itinerary />} />
+          <Route path="/profile" element= {<Profile />} />
         </Routes>
         <Footer />
       </div>
